+++ CHANGELOG.md
@@ -20,3 +20,8 @@ March 22nd, 2023
 
 - Added refresh delay configuration option.
 - Add question mark icon to replace directional icons when there isn't enough information to display accurate data.
+- Updated GPS alert handling.
+    - GPS over-speed alerts are now rounded to 2 decimal places.
+    - Added support for frozen GPS alerts.
+- The start.sh script placeholder is now created when the index pages loads, rather than the first time the "Start" button is pressed.
+- Refined permissions verification process.